pyproject.toml
[project]
name = "timage-forecasting"
version = "0.1.0"
description = "multi-modal time series forecasting with image data"
readme = "README.md"
license = {text = "MIT"}
authors = [
  {name = "minkeymouse", email = "minkey.chang@gmail.com"}
]
requires-python = ">=3.11,<3.14"

packages = [
  { include = "timage_forecasting", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
triton = ">=3.3.1,<4.0.0"
numpy = ">=2.3.1,<3.0.0"
pandas = ">=2.3.0,<3.0.0"
torch = ">=2.7.1"
torchvision = ">=0.22.1"
torchaudio = ">=2.7.1"
pytorch-forecasting = ">=1.4.0"
transformers = ">=4.53.0,<5.0.0"
timm = ">=1.0.16,<2.0.0"

# your git dependency
mamba-ssm = { git = "https://github.com/state-spaces/mamba.git", rev = "v2.2.4" }
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
